+++ tests/Checks/HasEventHandlerTest.php
@@ -1,0 +1,150 @@
+<?php
+
+/**
+ * Copyright (c) 2015-present Ganbaro Digital Ltd
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *
+ *   * Redistributions in binary form must reproduce the above copyright
+ *     notice, this list of conditions and the following disclaimer in
+ *     the documentation and/or other materials provided with the
+ *     distribution.
+ *
+ *   * Neither the names of the copyright holders nor the names of his
+ *     contributors may be used to endorse or promote products derived
+ *     from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *
+ * @category  Libraries
+ * @package   EventStream/Checks
+ * @author    Stuart Herbert <stuherbert@ganbarodigital.com>
+ * @copyright 2015-present Ganbaro Digital Ltd www.ganbarodigital.com
+ * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
+ * @link      http://code.ganbarodigital.com/php-event-stream
+ */
+
+namespace GanbaroDigital\EventStream\Checks;
+
+use GanbaroDigital\EventStream\Events\Event;
+use GanbaroDigital\EventStream\Streams\EventStream;
+use GanbaroDigital\EventStream\Streams\RegisterEventHandler;
+use PHPUnit_Framework_TestCase;
+use RuntimeException;
+use stdClass;
+
+class HasEventHandlerTest_DummyEvent implements Event { }
+
+/**
+ * @coversDefaultClass GanbaroDigital\EventStream\Checks\HasEventHandler
+ */
+class HasEventHandlerTest extends PHPUnit_Framework_TestCase
+{
+    /**
+     * @covers ::check
+     */
+    public function testCanCallStatically()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $stream = new EventStream;
+        $eventName = HasEventHandlerTest_DummyEvent::class;
+        $handler = function($event) { };
+
+        $this->assertFalse(HasEventHandler::check($stream, $eventName));
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        RegisterEventHandler::on($stream, $eventName, $handler);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertTrue(HasEventHandler::check($stream, $eventName));
+    }
+
+    /**
+     * @covers ::check
+     * @dataProvider provideInvalidEventNames
+     */
+    public function testReturnsFalseForInvalidEventNames($eventName)
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $stream = new EventStream;
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $actualResult = HasEventHandler::check($stream, $eventName);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertFalse($actualResult);
+    }
+
+    /**
+     * @covers ::check
+     * @dataProvider provideInvalidClassNames
+     * @expectedException GanbaroDigital\Reflection\Exceptions\E4xx_NoSuchClass
+     */
+    public function testThrowsExceptionWhenClassNameDoesNotExist($eventName)
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $stream = new EventStream;
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $actualResult = HasEventHandler::check($stream, $eventName);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertFalse($actualResult);
+    }
+
+    public function provideInvalidEventNames()
+    {
+        return [
+            [ null, false ],
+            [ [], false ],
+            [ true, false ],
+            [ false, false ],
+            [ 3.1415927, false ],
+            [ 0, false ],
+            [ 100, false ],
+            [ new stdClass, false ],
+        ];
+    }
+
+    public function provideInvalidClassNames()
+    {
+        return [
+            [ "hello, world!", false ],
+        ];
+    }
+}